+++ xarray/core/missing.py
@@ -317,9 +317,13 @@ def interp_na(
         if not is_scalar(max_gap):
             raise ValueError("max_gap must be a scalar.")
 
+        # TODO: benbovy - flexible indexes: update when CFTimeIndex (and DatetimeIndex?)
+        # has its own class inheriting from xarray.Index
         if (
-            dim in self.indexes
-            and isinstance(self.indexes[dim], (pd.DatetimeIndex, CFTimeIndex))
+            dim in self.xindexes
+            and isinstance(
+                self.xindexes[dim].to_pandas_index(), (pd.DatetimeIndex, CFTimeIndex)
+            )
             and use_coordinate
         ):
             # Convert to float